requirements.txt
Django>=3.2.12, <4.0
psycopg2-binary>=2.9.3, <3.0
django-polymorphic>=3.1.0, <4.0
django-autocomplete-light>=3.9.4, <4.0
django_admin_sortable2>=1.0.4, <2.0
djhacker>=0.1.2, <1.0
djangorestframework>=3.13.1, <4.0
django-rest-polymorphic>=0.1.9, <1.0